# specs/001-build-a-website/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: Domaine Vallot E-Commerce API
  description: API for boutique winery direct-to-consumer e-commerce website
  version: 1.0.0
  contact:
    name: Domaine Vallot
    email: info@domainevallot.com
servers:
  - url: https://domainevallot.com/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /products/{productId}:
    get:
      summary: Get product details
      description: Retrieve detailed information for a specific wine product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: lang
          in: query
          schema:
            type: string
            enum: [en, fr]
            default: en
      responses:
        '200':
          description: Product details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetail'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        vintage:
          type: integer
        varietal:
          type: string
        price_eur:
          type: number
          format: decimal
        stock_quantity:
          type: integer
        description:
          type: string
        tasting_notes:
          type: string
        food_pairing:
          type: string
        images:
          type: array
          items:
            type: object
            properties:
              url:
                type: string
                format: uri
              alt_text:
                type: string
              is_primary:
                type: boolean
        certifications:
          type: array
          items:
            type: string
            enum: [organic, biodynamic, vegan]
        slug:
          type: string
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/Product'
        - type: object
          properties:
            alcohol_content:
              type: number
              format: decimal
            volume_ml:
              type: integer
            region:
              type: string
            production_notes:
              type: string
            allergens:
              type: array
              items:
                type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object